requirements.txt
discord.py
requests
beautifulsoup4
apscheduler
flask
